requirements.txt
pandas
sqlalchemy
psycopg2-binary
python-dotenv
requests
pytz
